chore: Moved service/session-manager/session.registry.mts to core/server/session.registry.mts.
