--- packages/architectura/src/service/session-manager/_index.mts
@@ -1,1 +1,0 @@
-export * from './session.registry.mjs';
